+++ pages/route-picker/show/route.php
@@ -1,7 +1,7 @@
 <!doctype html>
 <html lang="en">
     <head>
-        <link rel="stylesheet" href="../../style.css"/>
+        <link rel="stylesheet" href="../../../style.css"/>
         <link rel="stylesheet" href="route.css"/>
         <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -12,7 +12,7 @@
     <body id="bootstrap-overrides">
         <div class="route-container">
             <div class="header-container">
-                <a href="../../index.php" class="back-btn">Pick another route</a>
+                <a href="../../../index.php" class="back-btn">Pick another route</a>
             </div>
             <div class="image-box">
                 <img id="image-id" src="" />
